Add static container build
Use a container with musl libc inside, which can yield Go-style static
binaries for Rust. This allows us to ship containers that contain only
the binary, and not an entire linux distro. We don't want the distro, we
just want the libc.

conatiner.sh runs the musl build inside the musl environment, and leaves
the binary on our local disk. The Dockerfile copies in the binary and
exposes our ports, and yields a small Docker image.

wip